requirements.txt
# requirements.txt for CyberSec Log Analyzer
clickhouse-connect
duckdb
pandas
xgboost
scikit-learn
streamlit
fastapi
uvicorn
openai
qdrant-client
pinecone-client
python-dotenv
